feat: Add create method with beforeSaveCallback to AbstractCrudBackendApiController
This commit adds a new overloaded `create` method to the `AbstractCrudBackendApiController` class. The new method accepts an additional parameter `beforeSaveCallback`, which is a function that can be used to modify the entity before saving it to the repository. This allows for more flexibility in customizing the creation process of entities.

Note: This commit message follows the established convention of using the "feat" type for new feature additions.