refactor: Add ingestors field to SiloEntity
This commit adds a new field `ingestors` to the `SiloEntity` class. The `ingestors` field is a list of Ingestor UUIDs that are linked to the Silo. When linked, the Ingestors will be run in order to add documents into the Silo. This change enhances the functionality of the SiloEntity by allowing it to manage and track the associated Ingestors.