Documents : Save them inside database (#106)
* Adding DocumentEntity

* Saving Documents to Database

* "Try" enhancements

* Ability to delete a DocumentEntity and its linked embeddings chunks

* Ability to disallow a file ingestion

* feat: Update SearchRagApiController to use SiloEntity in query method

This commit updates the `SearchRagApiController` class to use the `SiloEntity` object in the `query` method instead of just the UUID.

* feat: Add functionality to delete documents not seen since the start of the ingestion

This commit adds a new method `deleteDocumentsInSiloNotSeenSince` to the `IngestorService` class. This method deletes documents in a Silo that have not been seen since the start of the ingestion process. It helps to clean up the database by removing documents that are no longer relevant.

* feat: Clean document chunks before ingesting

This commit modifies the `IngestorService` class to clean the chunks of a document before ingesting it into the embedding store. This helps to ensure that only the latest and relevant data is stored in the embedding store. Additionally, it updates the `IngestorService` class to include a new method `deleteDocumentsInSiloNotSeenSince` for deleting documents in a Silo that have not been seen since the start of the ingestion process. These changes improve the efficiency and accuracy of the ingestion process and help to maintain a clean and relevant database.

* If document is not changed don’t embedded it again

* No .class should be there

* *.class files should not be on Git

* feat: Clean document chunks before ingesting