refactor: Update Try class to support fallback function
The Try class in Try.java has been updated to support a fallback function. This allows users to provide a function that will be executed if an exception is thrown during the execution of the Try block. The fallback function takes a Throwable as input and returns a value of type T. This update enhances the error handling capabilities of the Try class.